The ParticleFilter Module

ParticleFilter — Module

ParticleFilter.jl

This is the bones for a discrete-time particle filter that I built to infer the infection numbers based on wastewater surveillance data

Module Index

ParticleFilter.Filter

ParticleFilter.Particle

ParticleFilter.Particle

ParticleFilter.average_particle

ParticleFilter.compute_likelihood

ParticleFilter.max_weight

ParticleFilter.multinomial_sampling!

ParticleFilter.needs_resampling

ParticleFilter.normalize_weights!

ParticleFilter.percentile_sampling!

ParticleFilter.propogate_sample!

ParticleFilter.quant

Detailed API

ParticleFilter.Filter — Type

Particle Filter Struct

T - Time

particles - Vector of particles

Measurements - Array of Measurements

MeasurementModel - A function that takes a measurement and a particle and outputs a real number

DynamicModel - An inplace function that mutates a particle

ParticleFilter.Particle — Type

Custom Type for particles Particle(weight::Float64,state::Vector{:<Real},pars::NamedTuple)

ParticleFilter.Particle — Method

Method for constructing particles

ParticleFilter.average_particle — Method

averageparticle(ParticleFilter) Determines the average particle from given particle filter

ParticleFilter.compute_likelihood — Method

Calculate the likelihood using the measurement model

ParticleFilter.max_weight — Method

Return the particle with the highest weight

ParticleFilter.multinomial_sampling! — Method

Resample the particles

ParticleFilter.needs_resampling — Method

ParticleFilter.normalize_weights! — Method

Normalize the weights

ParticleFilter.percentile_sampling! — Method

percentile_sampling(particle_filter,measurement;q::Float64)

resamples the particles based on the current measurement. Particles with a weight in the lowest q-th percentile are dropped and replaced by particles with the highest percentile.

ParticleFilter.propogate_sample! — Method

propogatesample!(particlefilter,t)

Uses the dynamic model to push particles forward one time step.

ParticleFilter.quant — Method

quant(particle_filter,q::Float64) Determines the weight that separates the particles into two groups