The ParticleFilter Module

ParticleFilter — Module

ParticleFilter.jl

This is the bones for a discrete-time particle filter that I built to infer the infection numbers based on wastewater surveillance data

Module Index

ParticleFilter.Filter

ParticleFilter.Particle

ParticleFilter.average_particle

ParticleFilter.avg_pars

ParticleFilter.compute_likelihood

ParticleFilter.max_weight

ParticleFilter.multinomial_sampling!

ParticleFilter.normalize_weights!

ParticleFilter.percentile_sampling!

ParticleFilter.propogate_sample!

ParticleFilter.quant

ParticleFilter.update!

Detailed API

ParticleFilter.Filter — Type

Particle(T::Real,particles::Vector{Particle},Measurements::Array,MeasurementModel::Function,DynamicModel::Function)

T - Time

particles - Vector of particles

Measurements - Array of Measurements

MeasurementModel - A function that takes a measurement and a particle and outputs a real number

DynamicModel - An inplace function that mutates a particle

ParticleFilter.Particle — Type

Particle(weight::Float64,state::Vector{:<Real},pars::NamedTuple)

ParticleFilter.average_particle — Method

average_particle(p::Filter)

Determines the average particle from given particle filter

ParticleFilter.avg_pars — Method

avg_pars(p::Filter)

Find the average of the parameters

ParticleFilter.compute_likelihood — Method

compute_likelihood(p::filter,sample::particle)

Calculate the likelihood using the measurement model

ParticleFilter.max_weight — Method

max_weight(p::Filter)

Return the particle with the highest weight

ParticleFilter.multinomial_sampling! — Method

multinomial_sampling!(p::Filter)

Resamples particles using a multinomial sampling procedure. If n is the number of particles, then this draws n samples from a multinomial distribution defined using the weights of the particles

ParticleFilter.normalize_weights! — Method

normalize_weights(p::Filter)

Normalize the weights

ParticleFilter.percentile_sampling! — Method

percentile_sampling(particle_filter,measurement;q::Float64)

resamples the particles based on the current measurement. Particles with a weight in the lowest q-th percentile are dropped and replaced by particles with the highest percentile.

ParticleFilter.propogate_sample! — Method

propogate_sample!(p::Filter,sample::Particle)

Uses the dynamic model to push particles forward one time step.

ParticleFilter.quant — Method

quant(particle_filter,q::Float64)

Determines the weight that separates the particles into two groups

ParticleFilter.update! — Method

update!(p::Filter)